Fix zero hour
Prior to this change, both * and 0 were treated the same when in the
hour position. This is incorrect, as * means every hour of the day but 0
specifically means the 0th hour -- midnight, aka 0:00 aka 12:00 AM.

This fixes the parsing such that 0 will properly return what it actually
means -- only during the 12:00 AM time slot.